requirements.txt
load_dotenv
ollama
crewai
langchain_google_genai
langchain_openai